Reworked test handling -- again.
Turns out there were two issues with the tests.

Issue 1: My character counter was "off" when I needed to inject
characters into the stream, so I had to rewrite that to reset the index
on the start of a new packet.

Issue 2: The code would totally fail if the SOH signal was corrupted.
It would read through the next few bytes in the buffer looking for it
and then gag out.  So we just flush the input buffer and wait for the
sender to give us a new packet.  On the Linux system, it seems reliable.